update-entity

+++ PetRescue/PetRescue.Data/Models/Center.cs
@@ -14,7 +14,7 @@ namespace PetRescue.Data.Models
         public string CenterName { get; set; }
         public string Address { get; set; }
         public string Phone { get; set; }
-        public byte CenterStatus { get; set; }
+        public int CenterStatus { get; set; }
         public Guid InsertBy { get; set; }
         public DateTime InsertAt { get; set; }
         public Guid? UpdateBy { get; set; }
